analytics-service/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
textblob
scikit-learn
python-dotenv
pymongo